+++ src/capture/middlewares_list.c.in
@@ -1,5 +1,5 @@
 
-#include "middlewares.h"
+#include "middlewares_list.h"
 
 #cmakedefine _CAPTURE_MIDDLEWARES_DEFINITIONS @_CAPTURE_MIDDLEWARES_DEFINITIONS@
 #cmakedefine _CAPTURE_MIDDLEWARES_LIST_INITIALISER @_CAPTURE_MIDDLEWARES_LIST_INITIALISER@
